@@ -4,13 +4,13 @@
     "license": "MIT",
     "type": "library",
     "require": {
-        "php": "^7.1|^8.0",
-        "articus/data-transfer": "^0.5",
-        "psr/container": "^1.0",
+        "php": "^7.4|^8.0",
+        "articus/data-transfer": "^0.6",
+        "articus/plugin-manager": "^0.1.1",
+        "psr/container": "^1.0|^2.0",
         "psr/http-client": "^1.0",
         "psr/http-factory": "^1.0",
-        "psr/http-message": "^1.0",
-        "laminas/laminas-servicemanager": "^3.4"
+        "psr/http-message": "^1.0"
     },
     "autoload": {
         "psr-4": {
@@ -26,6 +26,7 @@
         "ext-uopz": "*",
         "kahlan/kahlan": "^5.1",
         "mockery/mockery": "^1.3",
-        "symfony/polyfill-mbstring": "^1.27"
+        "symfony/polyfill-mbstring": "^1.27",
+        "laminas/laminas-servicemanager": "^3.17"
     }
 }
