@@ -3,36 +3,29 @@ declare(strict_types=1);
 
 namespace OpenAPIGenerator\APIClient\BodyCoder;
 
+use InvalidArgumentException;
 use OpenAPIGenerator\APIClient\BodyCoderInterface;
 use Psr\Http\Message\StreamFactoryInterface;
 use Psr\Http\Message\StreamInterface;
+use function json_decode;
+use function json_encode;
+use function json_last_error;
+use function json_last_error_msg;
+use function sprintf;
+use const JSON_ERROR_NONE;
+use const JSON_OBJECT_AS_ARRAY;
 
 class Json implements BodyCoderInterface
 {
-	/**
-	 * @var StreamFactoryInterface
-	 */
-	protected $streamFactory;
+	protected StreamFactoryInterface $streamFactory;
 
-	/**
-	 * @var int
-	 */
-	protected $encodeFlags;
+	protected int $encodeFlags;
 
-	/**
-	 * @var int
-	 */
-	protected $decodeFlags;
+	protected int $decodeFlags;
 
-	/**
-	 * @var bool
-	 */
-	protected $decodeAsAssociativeArray;
+	protected bool $decodeAsAssociativeArray;
 
-	/**
-	 * @var int
-	 */
-	protected $depth;
+	protected int $depth;
 	/**
 	 * @param StreamFactoryInterface $streamFactory
 	 * @param int $encodeFlags
@@ -44,7 +37,7 @@ class Json implements BodyCoderInterface
 		$this->streamFactory = $streamFactory;
 		$this->encodeFlags = $encodeFlags;
 		$this->decodeFlags = $decodeFlags;
-		$this->decodeAsAssociativeArray = (bool)($this->decodeFlags & \JSON_OBJECT_AS_ARRAY);
+		$this->decodeAsAssociativeArray = (bool)($this->decodeFlags & JSON_OBJECT_AS_ARRAY);
 		$this->depth = $depth;
 	}
 
@@ -53,10 +46,10 @@ class Json implements BodyCoderInterface
 	 */
 	public function encode($data): StreamInterface
 	{
-		$json = \json_encode($data, $this->encodeFlags, $this->depth);
+		$json = json_encode($data, $this->encodeFlags, $this->depth);
 		if ($json === false)
 		{
-			throw new \InvalidArgumentException(\sprintf('JSON encoding failure. %s', \json_last_error_msg()));
+			throw new InvalidArgumentException(sprintf('JSON encoding failure. %s', json_last_error_msg()));
 		}
 		return $this->streamFactory->createStream($json);
 	}
@@ -66,10 +59,10 @@ class Json implements BodyCoderInterface
 	 */
 	public function decode(StreamInterface $body)
 	{
-		$result = \json_decode($body->getContents(), $this->decodeAsAssociativeArray, $this->depth, $this->decodeFlags);
-		if (($result === null) && (\json_last_error() !== \JSON_ERROR_NONE))
+		$result = json_decode($body->getContents(), $this->decodeAsAssociativeArray, $this->depth, $this->decodeFlags);
+		if (($result === null) && (json_last_error() !== JSON_ERROR_NONE))
 		{
-			throw new \InvalidArgumentException(\sprintf('JSON decoding failure. %s', \json_last_error_msg()));
+			throw new InvalidArgumentException(sprintf('JSON decoding failure. %s', json_last_error_msg()));
 		}
 		return $result;
 	}
