@@ -5,21 +5,23 @@ namespace spec\Utility;
 
 use Mockery\Expectation;
 use Mockery\MockInterface;
+use function array_keys;
+use function extension_loaded;
+use function mock;
+use function uopz_set_return;
+use function uopz_unset_return;
 
 class GlobalFunctionMock
 {
-	/**
-	 * @var MockInterface
-	 */
-	protected static $innerMock;
+	protected static ?MockInterface $innerMock = null;
 	/**
 	 * @var string[]
 	 */
-	protected static $functionNameMap = [];
+	protected static array $functionNameMap = [];
 
 	public static function disabled(): bool
 	{
-		return (!\extension_loaded('uopz'));
+		return (!extension_loaded('uopz'));
 	}
 
 	/**
@@ -30,17 +32,14 @@ class GlobalFunctionMock
 	{
 		if (!isset(self::$innerMock))
 		{
-			self::$innerMock = \mock();
+			self::$innerMock = mock();
 		}
 		if (!isset(self::$functionNameMap[$functionName]))
 		{
 			$mock = self::$innerMock;
-			\uopz_set_return(
+			uopz_set_return(
 				$functionName,
-				function(...$arguments) use ($mock, $functionName)
-				{
-					return $mock->{$functionName}(...$arguments);
-				},
+				static fn (...$arguments) => $mock->{$functionName}(...$arguments),
 				true
 			);
 			self::$functionNameMap[$functionName] = true;
@@ -50,9 +49,9 @@ class GlobalFunctionMock
 
 	public static function reset(): void
 	{
-		foreach (\array_keys(self::$functionNameMap) as $functionName)
+		foreach (array_keys(self::$functionNameMap) as $functionName)
 		{
-			\uopz_unset_return($functionName);
+			uopz_unset_return($functionName);
 			unset(self::$functionNameMap[$functionName]);
 		}
 		self::$innerMock = null;
